Move custom publication code to pipeline helpers
